// play/chess-tourney-frontend/src/api/model/leaderboardSortBy.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * (title)
 * API for a custom three-player chess variant game server. Authentication is required via a JWT Bearer token for protected endpoints.
 * OpenAPI spec version: 0.0.0
 */

export type LeaderboardSortBy =
  (typeof LeaderboardSortBy)[keyof typeof LeaderboardSortBy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const LeaderboardSortBy = {
  Wins: "Wins",
  GamesPlayed: "GamesPlayed",
  WinRate: "WinRate",
  LossRate: "LossRate",
  Losses: "Losses",
  Draws: "Draws",
} as const;